Category Model Added & Course method modification

+++ IMSProject/IMSapp/courseviews.py
@@ -2,6 +2,23 @@ from django.shortcuts import render,redirect,get_object_or_404
 from IMSapp.forms import *
 import os
 
+#------------Course Category---------
+def categorylist(request):
+    categorydata = CourseCategoryModel.objects.all()
+    
+    return render(request,'courses/categorylist.html',{'categorydata':categorydata})
+
+def addcategory(request):
+    if request.method == 'POST':
+        categoryform = CourseCategoryForm(request.POST)
+        if categoryform.is_valid():
+            categoryform.save()
+            return redirect('categorylist')
+    else:
+        categoryform = CourseCategoryForm()
+    
+    return render(request,'courses/addcategory.html',{'categoryform':categoryform})
+
 #------------Admin Courses-------------
 def addcourse(request):
     if request.method == 'POST':
@@ -26,12 +43,16 @@ def courselist(request):
 
 def editcourse(request,myid):
     coursedata = get_object_or_404(CourseInfoModel,id=myid)
+    img = coursedata.CourseImage
     if request.method == 'POST':
-        img = coursedata.CourseImage
-        os.remove(img.path)
         courseform = CourseInfoForm(request.POST, request.FILES, instance=coursedata)
         if courseform.is_valid():
-            courseform.save()
+            course = courseform.save(commit=False)
+            image = course.CourseImage
+            print("Image is: ",image)
+            if image != img:
+                os.remove(img.path)
+            course.save()
             return redirect('courselist')
     else:
         courseform = CourseInfoForm(instance=coursedata)
